+++ Sources/Overload/OvCore/include/OvCore/ECS/Components/Behaviour.inl
@@ -7,7 +7,7 @@
 #pragma once
 
 
-#include <OvDebug/Utils/Logger.h>
+#include <OvDebug/Logger.h>
 
 #include "OvCore/ECS/Components/Behaviour.h"
