+++ Sources/Overload/OvCore/include/OvCore/Resources/Material.inl
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include <OvDebug/Utils/Logger.h>
+#include <OvDebug/Logger.h>
 
 #include "OvCore/Resources/Material.h"
