+++ Sources/Overload/OvDebug/include/OvDebug/ILogHandler.h
