+++ Sources/Overload/OvDebug/src/OvDebug/Assertion.cpp
@@ -4,7 +4,7 @@
 * @licence: MIT
 */
 
-#include "OvDebug/Utils/Assertion.h"
+#include "OvDebug/Assertion.h"
 
 #include <assert.h>
