+++ Sources/Overload/OvDebug/src/OvDebug/FileHandler.cpp
@@ -8,7 +8,7 @@
 
 #include <OvTools/Time/Date.h>
 
-#include "OvDebug/Utils/FileHandler.h"
+#include "OvDebug/FileHandler.h"
 
 #include <fstream>
 #include <iostream>
