+++ Sources/Overload/OvRendering/src/OvRendering/Resources/Loaders/ShaderLoader.cpp
@@ -9,7 +9,7 @@
 
 #include <GL/glew.h>
 
-#include <OvDebug/Utils/Logger.h>
+#include <OvDebug/Logger.h>
 
 #include "OvRendering/Resources/Loaders/ShaderLoader.h"
